-Split UInt512 file.
