+++ src/MissingValues/UInt512.GenericMath.cs
@@ -1,12 +1,9 @@
 ﻿using MissingValues.Info;
 using MissingValues.Internals;
-using System.Buffers.Binary;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Numerics;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 using System.Runtime.Intrinsics;
 using MissingValues.Primitives;
 
@@ -111,60 +108,6 @@ namespace MissingValues
 		}
 		
 		static UInt512 IBigInteger<UInt512>.Create(ReadOnlySpan<ulong> parts) => new(parts);
-
-		/// <inheritdoc/>
-		public static UInt512 CreateChecked<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			UInt512 result;
-
-			if (value is UInt512 v)
-			{
-				result = v;
-			}
-			else if (!UInt512.TryConvertFromChecked(value, out result) && !TOther.TryConvertToChecked<UInt512>(value, out result))
-			{
-				Thrower.NotSupported<UInt512, TOther>();
-			}
-
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 CreateSaturating<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			UInt512 result;
-
-			if (value is UInt512 v)
-			{
-				result = v;
-			}
-			else if (!UInt512.TryConvertFromSaturating(value, out result) && !TOther.TryConvertToSaturating<UInt512>(value, out result))
-			{
-				Thrower.NotSupported<UInt512, TOther>();
-			}
-
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 CreateTruncating<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			UInt512 result;
-
-			if (value is UInt512 v)
-			{
-				result = v;
-			}
-			else if (!UInt512.TryConvertFromTruncating(value, out result) && !TOther.TryConvertToTruncating<UInt512>(value, out result))
-			{
-				Thrower.NotSupported<UInt512, TOther>();
-			}
-
-			return result;
-		}
 
 		/// <inheritdoc/>
 		public static (UInt512 Quotient, UInt512 Remainder) DivRem(UInt512 left, UInt512 right)
@@ -599,224 +542,6 @@ namespace MissingValues
 			return NumberFormatter.FormatUInt(in this, format, formatProvider);
 		}
 
-		static bool INumberBase<UInt512>.TryConvertFromChecked<TOther>(TOther value, out UInt512 result) => TryConvertFromChecked(value, out result);
-		private static bool TryConvertFromChecked<TOther>(TOther value, out UInt512 result)
-		{
-			bool converted = true;
-			checked
-			{
-				result = value switch
-				{
-					char actual => (UInt512)actual,
-					Half actual => (UInt512)actual,
-					float actual => (UInt512)actual,
-					double actual => (UInt512)actual,
-					Quad actual => (UInt512)actual,
-					decimal actual => (UInt512)actual,
-					byte actual => (UInt512)actual,
-					ushort actual => (UInt512)actual,
-					uint actual => (UInt512)actual,
-					ulong actual => (UInt512)actual,
-					UInt128 actual => (UInt512)actual,
-					UInt256 actual => (UInt512)actual,
-					UInt512 actual => actual,
-					nuint actual => (UInt512)actual,
-					sbyte actual => (UInt512)actual,
-					short actual => (UInt512)actual,
-					int actual => (UInt512)actual,
-					long actual => (UInt512)actual,
-					Int128 actual => (UInt512)actual,
-					Int256 actual => (UInt512)actual,
-					Int512 actual => (UInt512)actual,
-					BigInteger actual => (UInt512)actual,
-					_ => BitHelper.DefaultConvert<UInt512>(out converted)
-				};
-			}
-			return converted;
-		}
-
-		static bool INumberBase<UInt512>.TryConvertFromSaturating<TOther>(TOther value, out UInt512 result) => TryConvertFromSaturating(value, out result);
-		private static bool TryConvertFromSaturating<TOther>(TOther value, out UInt512 result)
-		{
-			const double TwoPow512 = 13407807929942597099574024998205846127479365820592393377723561443721764030073546976801874298166903427690031858186486050853753882811946569946433649006084096.0;
-
-			bool converted = true;
-			result = value switch
-			{
-				char actual => actual,
-				Half actual => (actual < Half.Zero) ? MinValue : (UInt512)actual,
-				float actual => (actual < 0) ? MinValue : (UInt512)actual,
-				double actual => (actual < 0) ? MinValue : (actual > TwoPow512) ? MaxValue : (UInt512)actual,
-				Quad actual => (actual >= new Quad(0x41FF_0000_0000_0000, 0x0000_0000_0000_0000)) ? UInt512.MaxValue : (actual <= Quad.Zero) ? UInt512.MinValue : (UInt512)actual,
-				decimal actual => (actual < 0) ? MinValue : (UInt512)actual,
-				byte actual => actual,
-				ushort actual => actual,
-				uint actual => actual,
-				ulong actual => actual,
-				UInt128 actual => actual,
-				UInt256 actual => actual,
-				UInt512 actual => actual,
-				nuint actual => actual,
-				sbyte actual => (actual < 0) ? MinValue : (UInt512)actual,
-				short actual => (actual < 0) ? MinValue : (UInt512)actual,
-				int actual => (actual < 0) ? MinValue : (UInt512)actual,
-				long actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int128 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int256 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int512 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				BigInteger actual => (BigInteger.IsNegative(actual)) ? MinValue : (actual > (BigInteger)MaxValue) ? MaxValue : (UInt512)actual,
-				_ => BitHelper.DefaultConvert<UInt512>(out converted)
-			};
-			return converted;
-		}
-
-		static bool INumberBase<UInt512>.TryConvertFromTruncating<TOther>(TOther value, out UInt512 result) => TryConvertFromTruncating(value, out result);
-		private static bool TryConvertFromTruncating<TOther>(TOther value, out UInt512 result)
-		{
-			bool converted = true;
-			result = value switch
-			{
-				char actual => actual,
-				Half actual => (actual < Half.Zero) ? MinValue : (UInt512)actual,
-				float actual => (actual < 0) ? MinValue : (UInt512)actual,
-				double actual => (actual < 0) ? MinValue : (UInt512)actual,
-				decimal actual => (actual < 0) ? MinValue : (UInt512)actual,
-				byte actual => actual,
-				ushort actual => actual,
-				uint actual => actual,
-				ulong actual => actual,
-				UInt128 actual => actual,
-				UInt256 actual => actual,
-				UInt512 actual => actual,
-				nuint actual => actual,
-				sbyte actual => (actual < 0) ? MinValue : (UInt512)actual,
-				short actual => (actual < 0) ? MinValue : (UInt512)actual,
-				int actual => (actual < 0) ? MinValue : (UInt512)actual,
-				long actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int128 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int256 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				Int512 actual => (actual < 0) ? MinValue : (UInt512)actual,
-				BigInteger actual => (BigInteger.IsNegative(actual)) ? MinValue : (UInt512)actual,
-				_ => BitHelper.DefaultConvert<UInt512>(out converted)
-			};
-			return converted;
-		}
-
-		static bool INumberBase<UInt512>.TryConvertToChecked<TOther>(UInt512 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-			checked
-			{
-				result = result switch
-				{
-					char => (TOther)(object)(char)value,
-					Half => (TOther)(object)(Half)value,
-					float => (TOther)(object)(float)value,
-					double => (TOther)(object)(double)value,
-					decimal => (TOther)(object)(decimal)value,
-					byte => (TOther)(object)(byte)value,
-					ushort => (TOther)(object)(ushort)value,
-					uint => (TOther)(object)(uint)value,
-					ulong => (TOther)(object)(ulong)value,
-					UInt128 => (TOther)(object)(UInt128)value,
-					UInt256 => (TOther)(object)(UInt256)value,
-					UInt512 => (TOther)(object)value,
-					nuint => (TOther)(object)(nuint)value,
-					sbyte => (TOther)(object)(sbyte)value,
-					short => (TOther)(object)(short)value,
-					int => (TOther)(object)(int)value,
-					long => (TOther)(object)(long)value,
-					Int128 => (TOther)(object)(Int128)value,
-					Int256 => (TOther)(object)(Int256)value,
-					Int512 => (TOther)(object)(Int512)value,
-					nint => (TOther)(object)(nint)value,
-					BigInteger => (TOther)(object)(BigInteger)value,
-					_ => BitHelper.DefaultConvert<TOther>(out converted)
-				};
-			}
-
-			return converted;
-		}
-
-		static bool INumberBase<UInt512>.TryConvertToSaturating<TOther>(UInt512 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-			result = result switch
-			{
-				char => (TOther)(object)(char)value,
-				Half => (TOther)(object)(Half)value,
-				float => (TOther)(object)(float)value,
-				double => (TOther)(object)(double)value,
-				decimal => (TOther)(object)(decimal)value,
-				byte => (TOther)(object)((value >= 0xFF) ? byte.MaxValue : (byte)value),
-				ushort => (TOther)(object)((value >= 0xFFFF) ? ushort.MaxValue : (ushort)value),
-				uint => (TOther)(object)((value >= 0xFFFF_FFFF) ? uint.MaxValue : (uint)value),
-				ulong => (TOther)(object)((value >= 0xFFFF_FFFF_FFFF_FFFF) ? ulong.MaxValue : (ulong)value),
-				UInt128 => (TOther)(object)((value >= new UInt256(0x0000_0000_0000_0000, 0x0000_0000_0000_0000, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF)) ? UInt128.MaxValue : (UInt128)value),
-				UInt256 => (TOther)(object)((value >= new UInt256(0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF)) ? UInt256.MaxValue : (UInt256)value),
-				UInt512 => (TOther)(object)value,
-#if TARGET_32BIT
-				nuint => (TOther)(object)((value >= 0x0000_0000_FFFF_FFFF) ? nuint.MaxValue : (nuint)value),
-#else
-				nuint => (TOther)(object)((value >= 0xFFFF_FFFF_FFFF_FFFF) ? nuint.MaxValue : (nuint)value),
-#endif
-				sbyte => (TOther)(object)((value >= 0x0000_0000_0000_007F) ? sbyte.MaxValue : (sbyte)value),
-				short => (TOther)(object)((value >= 0x0000_0000_0000_7FFF) ? short.MaxValue : (short)value),
-				int => (TOther)(object)((value >= 0x0000_0000_7FFF_FFFF) ? int.MaxValue : (int)value),
-				long => (TOther)(object)((value >= 0x7FFF_FFFF_FFFF_FFFF) ? long.MaxValue : (long)value),
-				Int128 => (TOther)(object)((value >= new UInt256(0x0000_0000_0000_0000, 0x0000_0000_0000_0000, 0x7FFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF)) ? Int128.MaxValue : (Int128)value),
-				Int256 => (TOther)(object)((value >= new UInt256(0x7FFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF)) ? Int256.MaxValue : (Int256)value),
-				Int512 => (TOther)(object)((value >= new UInt512(0x7FFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF)) ? Int512.MaxValue : (Int512)value),
-#if TARGET_32BIT
-				nint => (TOther)(object)((value >= 0x0000_0000_7FFF_FFFF) ? nint.MaxValue : (nint)value),
-#else
-				nint => (TOther)(object)((value >= 0x7FFF_FFFF_FFFF_FFFF) ? nint.MaxValue : (nint)value),
-#endif
-				BigInteger => (TOther)(object)(BigInteger)value,
-				_ => BitHelper.DefaultConvert<TOther>(out converted)
-			};
-
-			return converted;
-		}
-
-		static bool INumberBase<UInt512>.TryConvertToTruncating<TOther>(UInt512 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-			unchecked
-			{
-				result = result switch
-				{
-					char => (TOther)(object)(char)value,
-					Half => (TOther)(object)(Half)value,
-					float => (TOther)(object)(float)value,
-					double => (TOther)(object)(double)value,
-					decimal => (TOther)(object)(decimal)value,
-					byte => (TOther)(object)(byte)value,
-					ushort => (TOther)(object)(ushort)value,
-					uint => (TOther)(object)(uint)value,
-					ulong => (TOther)(object)(ulong)value,
-					UInt128 => (TOther)(object)(UInt128)value,
-					UInt256 => (TOther)(object)(UInt256)value,
-					UInt512 => (TOther)(object)value,
-					nuint => (TOther)(object)(nuint)value,
-					sbyte => (TOther)(object)(sbyte)value,
-					short => (TOther)(object)(short)value,
-					int => (TOther)(object)(int)value,
-					long => (TOther)(object)(long)value,
-					Int128 => (TOther)(object)(Int128)value,
-					Int256 => (TOther)(object)(Int256)value,
-					Int512 => (TOther)(object)(Int512)value,
-					nint => (TOther)(object)(nint)value,
-					BigInteger => (TOther)(object)(BigInteger)value,
-					_ => BitHelper.DefaultConvert<TOther>(out converted)
-				};
-			}
-			return converted;
-		}
-
 		/// <inheritdoc/>
 		public bool TryFormat(Span<char> destination, out int charsWritten, [StringSyntax(StringSyntaxAttribute.NumericFormat)] ReadOnlySpan<char> format, IFormatProvider? provider)
 		{
@@ -873,829 +598,5 @@ namespace MissingValues
 		static int IFormattableInteger<UInt512>.Log2Int32(in UInt512 value) => BitHelper.Log2(in value);
 		static int IFormattableInteger<UInt512>.LeadingZeroCountInt32(in UInt512 value) => BitHelper.LeadingZeroCount(in value);
 		static void IFormattableUnsignedInteger<UInt512>.ToDecChars<TChar>(in UInt512 number, Span<TChar> destination, int digits) => NumberFormatter.UInt512ToDecChars(number, destination, digits);
-
-		/// <inheritdoc/>
-		public static UInt512 operator +(in UInt512 value) => value;
-
-		/// <inheritdoc/>
-		public static UInt512 operator +(in UInt512 left, in UInt512 right)
-		{
-			// For unsigned addition, we can detect overflow by checking `(x + y) < x`
-			// This gives us the carry to add to upper to compute the correct result
-
-			ulong part0 = left._p0 + right._p0;
-			ulong carry = (part0 < left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 + right._p1 + carry;
-			carry = (part1 < left._p1 || (carry == 1 && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 + right._p2 + carry;
-			carry = (part2 < left._p2 || (carry == 1 && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 + right._p3 + carry;
-			carry = (part3 < left._p3 || (carry == 1 && part3 == left._p3)) ? 1UL : 0UL;
-
-			ulong part4 = left._p4 + right._p4 + carry;
-			carry = (part4 < left._p4 || (carry == 1 && part4 == left._p4)) ? 1UL : 0UL;
-
-			ulong part5 = left._p5 + right._p5 + carry;
-			carry = (part5 < left._p5 || (carry == 1 && part5 == left._p5)) ? 1UL : 0UL;
-
-			ulong part6 = left._p6 + right._p6 + carry;
-			carry = (part6 < left._p6 || (carry == 1 && part6 == left._p6)) ? 1UL : 0UL;
-
-			ulong part7 = left._p7 + right._p7 + carry;
-
-			return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked +(in UInt512 left, in UInt512 right)
-		{
-			ulong part0 = left._p0 + right._p0;
-			ulong carry = (part0 < left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 + right._p1 + carry;
-			carry = (part1 < left._p1 || (carry == 1 && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 + right._p2 + carry;
-			carry = (part2 < left._p2 || (carry == 1 && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 + right._p3 + carry;
-			carry = (part3 < left._p3 || (carry == 1 && part3 == left._p3)) ? 1UL : 0UL;
-
-			ulong part4 = left._p4 + right._p4 + carry;
-			carry = (part4 < left._p4 || (carry == 1 && part4 == left._p4)) ? 1UL : 0UL;
-
-			ulong part5 = left._p5 + right._p5 + carry;
-			carry = (part5 < left._p5 || (carry == 1 && part5 == left._p5)) ? 1UL : 0UL;
-
-			ulong part6 = left._p6 + right._p6 + carry;
-			carry = (part6 < left._p6 || (carry == 1 && part6 == left._p6)) ? 1UL : 0UL;
-
-			ulong part7 = checked(left._p7 + right._p7 + carry);
-
-			return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator -(in UInt512 value) => Zero - value;
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked -(in UInt512 value) => checked(Zero - value);
-
-		/// <inheritdoc/>
-		public static UInt512 operator -(in UInt512 left, in UInt512 right)
-		{
-			// For unsigned subtract, we can detect overflow by checking `(x - y) > x`
-			// This gives us the borrow to subtract from upper to compute the correct result
-
-			ulong part0 = left._p0 - right._p0;
-			ulong borrow = (part0 > left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 - right._p1 - borrow;
-			borrow = (part1 > left._p1 || (borrow == 1UL && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 - right._p2 - borrow;
-			borrow = (part2 > left._p2 || (borrow == 1UL && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 - right._p3 - borrow;
-			borrow = (part3 > left._p3 || (borrow == 1UL && part3 == left._p3)) ? 1UL : 0UL;
-
-			ulong part4 = left._p4 - right._p4 - borrow;
-			borrow = (part4 > left._p4 || (borrow == 1UL && part4 == left._p4)) ? 1UL : 0UL;
-
-			ulong part5 = left._p5 - right._p5 - borrow;
-			borrow = (part5 > left._p5 || (borrow == 1UL && part5 == left._p5)) ? 1UL : 0UL;
-
-			ulong part6 = left._p6 - right._p6 - borrow;
-			borrow = (part6 > left._p6 || (borrow == 1UL && part6 == left._p6)) ? 1UL : 0UL;
-
-			ulong part7 = left._p7 - right._p7 - borrow;
-
-			return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked -(in UInt512 left, in UInt512 right)
-		{
-			// For unsigned subtract, we can detect overflow by checking `(x - y) > x`
-			// This gives us the borrow to subtract from upper to compute the correct result
-
-			ulong part0 = left._p0 - right._p0;
-			ulong borrow = (part0 > left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 - right._p1 - borrow;
-			borrow = (part1 > left._p1 || (borrow == 1UL && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 - right._p2 - borrow;
-			borrow = (part2 > left._p2 || (borrow == 1UL && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 - right._p3 - borrow;
-			borrow = (part3 > left._p3 || (borrow == 1UL && part3 == left._p3)) ? 1UL : 0UL;
-
-			ulong part4 = left._p4 - right._p4 - borrow;
-			borrow = (part4 > left._p4 || (borrow == 1UL && part4 == left._p4)) ? 1UL : 0UL;
-
-			ulong part5 = left._p5 - right._p5 - borrow;
-			borrow = (part5 > left._p5 || (borrow == 1UL && part5 == left._p5)) ? 1UL : 0UL;
-
-			ulong part6 = left._p6 - right._p6 - borrow;
-			borrow = (part6 > left._p6 || (borrow == 1UL && part6 == left._p6)) ? 1UL : 0UL;
-
-			ulong part7 = checked(left._p7 - right._p7 - borrow);
-
-			return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator ~(in UInt512 value)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v = Unsafe.BitCast<UInt512, Vector512<ulong>>(value);
-				var result = ~v;
-				return Unsafe.BitCast<Vector512<ulong>, UInt512>(result);
-			}
-			else
-			{
-				return new(~value._p7, ~value._p6, ~value._p5, ~value._p4, ~value._p3, ~value._p2, ~value._p1, ~value._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator ++(in UInt512 value) => value + One;
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked ++(in UInt512 value) => checked(value + One);
-
-		/// <inheritdoc/>
-		public static UInt512 operator --(in UInt512 value) => value - One;
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked --(in UInt512 value) => checked(value - One);
-
-		/// <inheritdoc/>
-		public static UInt512 operator *(in UInt512 left, in UInt512 right)
-		{
-			ulong up, low;
-
-			if (right._p7 == 0 && right._p6 == 0 && right._p5 == 0 && right._p4 == 0)
-			{
-				if (right._p3 == 0 && right._p2 == 0 && right._p1 == 0)
-				{
-					if (left._p7 == 0 && left._p6 == 0 && left._p5 == 0 && left._p4 == 0 && left._p3 == 0 && left._p2 == 0 && left._p1 == 0)
-					{
-						up = Calculator.BigMul(left._p0, right._p0, out low);
-						return new UInt512(0, 0, 0, 0, 0, 0, up, low);
-					}
-
-					return Calculator.Multiply(in left, right._p0, out _);
-				}
-				if (left._p7 == 0 && left._p6 == 0 && left._p5 == 0 && left._p4 == 0)
-				{
-					if (left._p3 == 0 && left._p2 == 0 && left._p1 == 0)
-					{
-						var temp = Calculator.Multiply(right.Lower, left._p0, out low);
-						return new UInt512(0, 0, 0, low, temp.Part3, temp.Part2, temp.Part1, temp.Part0);
-					}
-
-					return MathQ.BigMul(left.Lower, right.Lower);
-				}
-			}
-			else if (left._p7 == 0 && left._p6 == 0 && left._p5 == 0 && left._p4 == 0 && left._p3 == 0 && left._p2 == 0 && left._p1 == 0)
-			{
-				return Calculator.Multiply(in right, left._p0, out _);
-			}
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p0, 0);
-			ulong p0 = low;
-			(up, low) = Calculator.BigMulAdd(left._p1, right._p0, up);
-			ulong p1 = low;
-			(up, low) = Calculator.BigMulAdd(left._p2, right._p0, up);
-			ulong p2 = low;
-			(up, low) = Calculator.BigMulAdd(left._p3, right._p0, up);
-			ulong p3 = low;
-			(up, low) = Calculator.BigMulAdd(left._p4, right._p0, up);
-			ulong p4 = low;
-			(up, low) = Calculator.BigMulAdd(left._p5, right._p0, up);
-			ulong p5 = low;
-			(up, low) = Calculator.BigMulAdd(left._p6, right._p0, up);
-			ulong p6 = low;
-			(_, low) = Calculator.BigMulAdd(left._p7, right._p0, up);
-			ulong p7 = low;
-
-        	(up, low) = Calculator.BigMulAdd(left._p0, right._p1, 0);
-	        p1 = Calculator.AddWithCarry(p1, low, out ulong carry);
-	        up += carry;
-        	(up, low) = Calculator.BigMulAdd(left._p1, right._p1, up);
-	        p2 = Calculator.AddWithCarry(p2, low, out carry);
-	        up += carry;
-        	(up, low) = Calculator.BigMulAdd(left._p2, right._p1, up);
-        	p3 = Calculator.AddWithCarry(p3, low, out carry);
-			up += carry;
-        	(up, low) = Calculator.BigMulAdd(left._p3, right._p1, up);
-        	p4 = Calculator.AddWithCarry(p4, low, out carry);
-	        up += carry;
-        	(up, low) = Calculator.BigMulAdd(left._p4, right._p1, up);
-        	p5 = Calculator.AddWithCarry(p5, low, out carry);
-	        up += carry;
-        	(up, low) = Calculator.BigMulAdd(left._p5, right._p1, up);
-        	p6 = Calculator.AddWithCarry(p6, low, out carry);
-	        up += carry;
-        	(_, low) = Calculator.BigMulAdd(left._p6, right._p1, up);
-        	p7 += low;
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p2, 0);
-			p2 = Calculator.AddWithCarry(p2, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p1, right._p2, up);
-			p3 = Calculator.AddWithCarry(p3, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p2, right._p2, up);
-			p4 = Calculator.AddWithCarry(p4, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p3, right._p2, up);
-			p5 = Calculator.AddWithCarry(p5, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p4, right._p2, up);
-			p6 = Calculator.AddWithCarry(p6, low, out carry);
-			up += carry;
-			(_, low) = Calculator.BigMulAdd(left._p5, right._p2, up);
-			p7 += low;
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p3, 0);
-			p3 = Calculator.AddWithCarry(p3, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p1, right._p3, up);
-			p4 = Calculator.AddWithCarry(p4, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p2, right._p3, up);
-			p5 = Calculator.AddWithCarry(p5, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p3, right._p3, up);
-			p6 = Calculator.AddWithCarry(p6, low, out carry);
-			up += carry;
-			(_, low) = Calculator.BigMulAdd(left._p4, right._p3, up);
-			p7 += low;
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p4, 0);
-			p4 = Calculator.AddWithCarry(p4, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p1, right._p4, up);
-			p5 = Calculator.AddWithCarry(p5, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p2, right._p4, up);
-			p6 = Calculator.AddWithCarry(p6, low, out carry);
-			up += carry;
-			(_, low) = Calculator.BigMulAdd(left._p3, right._p4, up);
-			p7 += low;
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p5, 0);
-			p5 = Calculator.AddWithCarry(p5, low, out carry);
-			up += carry;
-			(up, low) = Calculator.BigMulAdd(left._p1, right._p5, up);
-			p6 = Calculator.AddWithCarry(p6, low, out carry);
-			up += carry;
-			(_, low) = Calculator.BigMulAdd(left._p2, right._p5, up);
-			p7 += low;
-
-			(up, low) = Calculator.BigMulAdd(left._p0, right._p6, 0);
-			p6 = Calculator.AddWithCarry(p6, low, out carry);
-			up += carry;
-			(_, low) = Calculator.BigMulAdd(left._p1, right._p6, up);
-			p7 += low;
-
-			(_, low) = Calculator.BigMulAdd(left._p0, right._p7, 0);
-			p7 += low;
-
-			return new UInt512(p7, p6, p5, p4, p3, p2, p1, p0);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked *(in UInt512 left, in UInt512 right)
-		{
-			UInt512 upper = BigMul(left, right, out UInt512 lower);
-
-			if (upper != Zero)
-			{
-				Thrower.ArithmeticOverflow(Thrower.ArithmeticOperation.Multiplication);
-			}
-
-			return lower;
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator /(in UInt512 left, in UInt512 right)
-		{
-			const int UIntCount = Size / sizeof(ulong);
-
-			if (right._p7 == 0 && right._p6 == 0 && right._p5 == 0 && right._p4 == 0)
-			{
-				if (right._p3 == 0 && right._p2 == 0 && right._p1 == 0)
-				{
-					if (right._p0 == 0)
-					{
-						Thrower.DivideByZero();
-					}
-					return Calculator.Divide(in left, right._p0);
-				}
-			}
-
-			if (right >= left)
-			{
-				return (right == left) ? One : Zero;
-			}
-
-			Span<ulong> quotientSpan = stackalloc ulong[UIntCount];
-			BitHelper.Write(quotientSpan, in left);
-
-			Span<ulong> divisorSpan = stackalloc ulong[UIntCount];
-			BitHelper.Write(divisorSpan, in right);
-
-			Span<ulong> rawBits = stackalloc ulong[UIntCount];
-			rawBits.Clear();
-
-			Calculator.Divide(
-				quotientSpan[..BitHelper.GetTrimLength(in left)],
-				divisorSpan[..BitHelper.GetTrimLength(in right)],
-				rawBits);
-
-			return new UInt512(rawBits);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator checked /(in UInt512 left, in UInt512 right) => left / right;
-
-		/// <inheritdoc/>
-		public static UInt512 operator %(in UInt512 left, in UInt512 right)
-		{
-			const int UIntCount = Size / sizeof(ulong);
-
-			if (right._p7 == 0 && right._p6 == 0 && right._p5 == 0 && right._p4 == 0)
-			{
-				if (right._p3 == 0 && right._p2 == 0 && right._p1 == 0)
-				{
-					if (right._p0 == 0)
-					{
-						Thrower.DivideByZero();
-					}
-					return Calculator.Remainder(in left, right._p0);
-				}
-			}
-
-			if (right == left)
-			{
-				return Zero;
-			}
-
-			if (right > left)
-			{
-				return left;
-			}
-
-			Span<ulong> quotientSpan = stackalloc ulong[UIntCount];
-			BitHelper.Write(quotientSpan, in left);
-
-			Span<ulong> divisorSpan = stackalloc ulong[UIntCount];
-			BitHelper.Write(divisorSpan, in right);
-
-			Span<ulong> rawBits = stackalloc ulong[UIntCount];
-			rawBits.Clear();
-
-			Calculator.Remainder(
-				quotientSpan[..BitHelper.GetTrimLength(in left)],
-				divisorSpan[..BitHelper.GetTrimLength(in right)],
-				rawBits);
-
-			return new UInt512(rawBits);
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator &(in UInt512 left, in UInt512 right)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<UInt512, Vector512<ulong>>(left);
-				var v2 = Unsafe.BitCast<UInt512, Vector512<ulong>>(right);
-				var result = v1 & v2;
-				return Unsafe.BitCast<Vector512<ulong>, UInt512>(result);
-			}
-			else
-			{
-				return new(left._p7 & right._p7, left._p6 & right._p6, left._p5 & right._p5, left._p4 & right._p4, left._p3 & right._p3, left._p2 & right._p2, left._p1 & right._p1, left._p0 & right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator |(in UInt512 left, in UInt512 right)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<UInt512, Vector512<ulong>>(left);
-				var v2 = Unsafe.BitCast<UInt512, Vector512<ulong>>(right);
-				var result = v1 | v2;
-				return Unsafe.BitCast<Vector512<ulong>, UInt512>(result);
-			}
-			else
-			{
-				return new(left._p7 | right._p7, left._p6 | right._p6, left._p5 | right._p5, left._p4 | right._p4, left._p3 | right._p3, left._p2 | right._p2, left._p1 | right._p1, left._p0 | right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator ^(in UInt512 left, in UInt512 right)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<UInt512, Vector512<ulong>>(left);
-				var v2 = Unsafe.BitCast<UInt512, Vector512<ulong>>(right);
-				var result = v1 ^ v2;
-				return Unsafe.BitCast<Vector512<ulong>, UInt512>(result);
-			}
-			else
-			{
-				return new(left._p7 ^ right._p7, left._p6 ^ right._p6, left._p5 ^ right._p5, left._p4 ^ right._p4, left._p3 ^ right._p3, left._p2 ^ right._p2, left._p1 ^ right._p1, left._p0 ^ right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator <<(in UInt512 value, int shiftAmount)
-		{
-			// C# automatically masks the shift amount for UInt64 to be 0x3F. So we
-			// need to specially handle things if the shift amount exceeds 0x3F.
-
-			shiftAmount &= 0x1FF; // mask the shift amount to be within [0, 255]
-
-			if (shiftAmount == 0)
-			{
-				return value;
-			}
-
-			if (shiftAmount < 64)
-			{
-				ulong part7 = (value._p7 << shiftAmount) | (value._p6 >> (64 - shiftAmount));
-				ulong part6 = (value._p6 << shiftAmount) | (value._p5 >> (64 - shiftAmount));
-				ulong part5 = (value._p5 << shiftAmount) | (value._p4 >> (64 - shiftAmount));
-				ulong part4 = (value._p4 << shiftAmount) | (value._p3 >> (64 - shiftAmount));
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 128)
-			{
-				shiftAmount -= 64;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p6, value._p5, value._p4, value._p3, value._p2, value._p1, value._p0, 0);
-				}
-
-				ulong part6 = (value._p6 << shiftAmount) | (value._p5 >> (64 - shiftAmount));
-				ulong part5 = (value._p5 << shiftAmount) | (value._p4 >> (64 - shiftAmount));
-				ulong part4 = (value._p4 << shiftAmount) | (value._p3 >> (64 - shiftAmount));
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part6, part5, part4, part3, part2, part1, part0, 0);
-			}
-			else if (shiftAmount < 192)
-			{
-				shiftAmount -= 128;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p5, value._p4, value._p3, value._p2, value._p1, value._p0, 0, 0);
-				}
-
-				ulong part5 = (value._p5 << shiftAmount) | (value._p4 >> (64 - shiftAmount));
-				ulong part4 = (value._p4 << shiftAmount) | (value._p3 >> (64 - shiftAmount));
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part5, part4, part3, part2, part1, part0, 0, 0);
-			}
-			else if (shiftAmount < 256)
-			{
-				shiftAmount -= 192;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p4, value._p3, value._p2, value._p1, value._p0, 0, 0, 0);
-				}
-
-				ulong part4 = (value._p4 << shiftAmount) | (value._p3 >> (64 - shiftAmount));
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part4, part3, part2, part1, part0, 0, 0, 0);
-			}
-			else if (shiftAmount < 320)
-			{
-				shiftAmount -= 256;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p3, value._p2, value._p1, value._p0, 0, 0, 0, 0);
-				}
-
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part3, part2, part1, part0, 0, 0, 0, 0);
-			}
-			else if (shiftAmount < 384)
-			{
-				shiftAmount -= 320;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p2, value._p1, value._p0, 0, 0, 0, 0, 0);
-				}
-
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part2, part1, part0, 0, 0, 0, 0, 0);
-			}
-			else if (shiftAmount < 448)
-			{
-				shiftAmount -= 384;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p1, value._p0, 0, 0, 0, 0, 0, 0);
-				}
-
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part1, part0, 0, 0, 0, 0, 0, 0);
-			}
-			else // shiftAmount < 512
-			{
-				shiftAmount -= 448;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(value._p0, 0, 0, 0, 0, 0, 0, 0);
-				}
-
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new UInt512(part0, 0, 0, 0, 0, 0, 0, 0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator >>(in UInt512 value, int shiftAmount) => value >>> shiftAmount;
-
-		/// <inheritdoc/>
-		public static bool operator ==(in UInt512 left, in UInt512 right)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v1 = Vector512.Create(left._p0, left._p1, left._p2, left._p3, left._p4, left._p5, left._p6, left._p7);
-				var v2 = Vector512.Create(right._p0, right._p1, right._p2, right._p3, right._p4, right._p5, right._p6, right._p7);
-				return v1 == v2;
-			}
-			if (Vector256.IsHardwareAccelerated)
-			{
-				Vector256<ulong> vUpper1 = Vector256.Create(left._p4, left._p5, left._p6, left._p7);
-				Vector256<ulong> vLower1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3);
-				
-				Vector256<ulong> vUpper2 = Vector256.Create(right._p4, right._p5, right._p6, right._p7);
-				Vector256<ulong> vLower2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3);
-
-				return vUpper1 == vUpper2 && vLower1 == vLower2;
-			}
-
-			return (left._p7 == right._p7) && (left._p6 == right._p6) && (left._p5 == right._p5) && (left._p4 == right._p4)
-			       && (left._p3 == right._p3) && (left._p2 == right._p2) && (left._p1 == right._p1) && (left._p0 == right._p0);
-		}
-
-		/// <inheritdoc/>
-		public static bool operator !=(in UInt512 left, in UInt512 right)
-		{
-			if (Vector512.IsHardwareAccelerated)
-			{
-				var v1 = Vector512.Create(left._p0, left._p1, left._p2, left._p3, left._p4, left._p5, left._p6, left._p7);
-				var v2 = Vector512.Create(right._p0, right._p1, right._p2, right._p3, right._p4, right._p5, right._p6, right._p7);
-				return v1 != v2;
-			}
-			if (Vector256.IsHardwareAccelerated)
-			{
-				Vector256<ulong> vUpper1 = Vector256.Create(left._p4, left._p5, left._p6, left._p7);
-				Vector256<ulong> vLower1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3);
-				
-				Vector256<ulong> vUpper2 = Vector256.Create(right._p4, right._p5, right._p6, right._p7);
-				Vector256<ulong> vLower2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3);
-
-				return vUpper1 != vUpper2 || vLower1 != vLower2;
-			}
-
-			return (left._p7 != right._p7) || (left._p6 != right._p6) || (left._p5 != right._p5) || (left._p4 != right._p4)
-			       || (left._p3 != right._p3) || (left._p2 != right._p2) || (left._p1 != right._p1) || (left._p0 != right._p0);
-		}
-
-		/// <inheritdoc/>
-		public static bool operator <(in UInt512 left, in UInt512 right)
-		{
-			// Successively compare each part.
-			return (left._p7 < right._p7)
-				|| (left._p7 == right._p7 && ((left._p6 < right._p6)
-				|| (left._p6 == right._p6 && ((left._p5 < right._p5)
-				|| (left._p5 == right._p5 && ((left._p4 < right._p4)
-				|| (left._p4 == right._p4 && ((left._p3 < right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 < right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 < right._p1)
-				|| (left._p1 == right._p1 && (left._p0 < right._p0))))))))))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator >(in UInt512 left, in UInt512 right)
-		{
-			return (left._p7 > right._p7)
-				|| (left._p7 == right._p7 && ((left._p6 > right._p6)
-				|| (left._p6 == right._p6 && ((left._p5 > right._p5)
-				|| (left._p5 == right._p5 && ((left._p4 > right._p4)
-				|| (left._p4 == right._p4 && ((left._p3 > right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 > right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 > right._p1)
-				|| (left._p1 == right._p1 && (left._p0 > right._p0))))))))))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator <=(in UInt512 left, in UInt512 right)
-		{
-			return (left._p7 < right._p7)
-				|| (left._p7 == right._p7 && ((left._p6 < right._p6)
-				|| (left._p6 == right._p6 && ((left._p5 < right._p5)
-				|| (left._p5 == right._p5 && ((left._p4 < right._p4)
-				|| (left._p4 == right._p4 && ((left._p3 < right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 < right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 < right._p1)
-				|| (left._p1 == right._p1 && (left._p0 <= right._p0))))))))))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator >=(in UInt512 left, in UInt512 right)
-		{
-			return (left._p7 > right._p7)
-				|| (left._p7 == right._p7 && ((left._p6 > right._p6)
-				|| (left._p6 == right._p6 && ((left._p5 > right._p5)
-				|| (left._p5 == right._p5 && ((left._p4 > right._p4)
-				|| (left._p4 == right._p4 && ((left._p3 > right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 > right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 > right._p1)
-				|| (left._p1 == right._p1 && (left._p0 >= right._p0))))))))))))));
-		}
-
-		/// <inheritdoc/>
-		public static UInt512 operator >>>(in UInt512 value, int shiftAmount)
-		{
-			// C# automatically masks the shift amount for UInt64 to be 0x3F. So we
-			// need to specially handle things if the shift amount exceeds 0x3F.
-
-			shiftAmount &= 0x1FF; // mask the shift amount to be within [0, 511]
-
-			if (shiftAmount == 0)
-			{
-				return value;
-			}
-
-			if (shiftAmount < 64)
-			{
-				ulong part0 = (value._p0 >> shiftAmount) | (value._p1 << (64 - shiftAmount));
-				ulong part1 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part2 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part3 = (value._p3 >> shiftAmount) | (value._p4 << (64 - shiftAmount));
-				ulong part4 = (value._p4 >> shiftAmount) | (value._p5 << (64 - shiftAmount));
-				ulong part5 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part6 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part7 = value._p7 >> shiftAmount;
-
-				return new UInt512(part7, part6, part5, part4, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 128)
-			{
-				shiftAmount -= 64;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, value._p7, value._p6, value._p5, value._p4, value._p3, value._p2, value._p1);
-				}
-
-				ulong part0 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part1 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part2 = (value._p3 >> shiftAmount) | (value._p4 << (64 - shiftAmount));
-				ulong part3 = (value._p4 >> shiftAmount) | (value._p5 << (64 - shiftAmount));
-				ulong part4 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part5 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part6 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, part6, part5, part4, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 192)
-			{
-				shiftAmount -= 128;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, 0, value._p7, value._p6, value._p5, value._p4, value._p3, value._p2);
-				}
-
-				ulong part0 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part1 = (value._p3 >> shiftAmount) | (value._p4 << (64 - shiftAmount));
-				ulong part2 = (value._p4 >> shiftAmount) | (value._p5 << (64 - shiftAmount));
-				ulong part3 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part4 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part5 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, part5, part4, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 256)
-			{
-				shiftAmount -= 192;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, 0, 0, value._p7, value._p6, value._p5, value._p4, value._p3);
-				}
-
-				ulong part0 = (value._p3 >> shiftAmount) | (value._p4 << (64 - shiftAmount));
-				ulong part1 = (value._p4 >> shiftAmount) | (value._p5 << (64 - shiftAmount));
-				ulong part2 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part3 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part4 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, 0, part4, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 320)
-			{
-				shiftAmount -= 256;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, 0, 0, 0, value._p7, value._p6, value._p5, value._p4);
-				}
-
-				ulong part0 = (value._p4 >> shiftAmount) | (value._p5 << (64 - shiftAmount));
-				ulong part1 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part2 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part3 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, 0, 0, part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 384)
-			{
-				shiftAmount -= 320;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, 0, 0, 0, 0, value._p7, value._p6, value._p5);
-				}
-
-				ulong part0 = (value._p5 >> shiftAmount) | (value._p6 << (64 - shiftAmount));
-				ulong part1 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part2 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, 0, 0, 0, part2, part1, part0);
-			}
-			else if (shiftAmount < 448)
-			{
-				shiftAmount -= 384;
-
-				if (shiftAmount == 0)
-				{
-					return new UInt512(0, 0, 0, 0, 0, 0, value._p7, value._p6);
-				}
-
-				ulong part0 = (value._p6 >> shiftAmount) | (value._p7 << (64 - shiftAmount));
-				ulong part1 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, 0, 0, 0, 0, part1, part0);
-			}
-			else // shiftAmount < 512
-			{
-				shiftAmount -= 448;
-
-				ulong part0 = value._p7 >> shiftAmount;
-
-				return new UInt512(0, 0, 0, 0, 0, 0, 0, part0);
-			}
-		}
 	}
 }
